requirements.txt
streamlit
python-dotenv
openai>=1.40
crewai
crewai-tools
pandas
tabulate
pydantic
requests
